pyproject.toml
[project]
name = "bmo-orq"
version = "0.2.0"
description = "Conference-booth fork of be-more-agent: Pi as I/O, orq.ai Agent as brain."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "Arian Pasquali" }]

dependencies = [
    "orq-ai-sdk>=0.5.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "Pillow>=10.0.0",
    "openwakeword>=0.6.0",
    "onnxruntime>=1.17.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "pyright>=1.1.380",
    "pre-commit>=3.7.0",
]
pi = [
    "picamera2; sys_platform == 'linux'",
]

[project.scripts]
bmo = "bmo.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bmo"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = []

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
include = ["src", "tests", "scripts"]
strict = ["src/bmo"]
pythonVersion = "3.11"
reportMissingTypeStubs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "live: hits real orq.ai workspace (requires ORQ_API_KEY in environment)",
]
addopts = "-m 'not live'"

[tool.coverage.run]
source = ["src/bmo"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false

# python-prctl is a transitive dep of picamera2; it has no macOS wheel
# so we skip building it from source when not on Linux.
[tool.uv]
no-build-package = ["python-prctl"]

